+++ CHANGELOG.md
@@ -48,6 +48,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- SMARTS pattern for `Hydrophobic` were not matching as expected due to the presence of
+  explicit hydrogens in the molecules. The SMARTS have been rewritten to account for
+  this.
 - ProLIF can now use the segment number as a `ResidueId` chain. This should help with
   solvated systems that may reuse the same resname, resnumber and chain and were using
   the segment number/identifier to distinguish them in the topology file.
@@ -59,6 +62,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- `Hydrophobic` SMARTS pattern now excludes all carbon linked to
+  nitrogen/oxygen/fluoride from being hydrophobic, previous versions were allowing such
+  carbons if they were aromatic.
 - Strip whitespaces in `ResidueId` name and chain.
 - Improved parsing for less standard residue names in `ResidueId`.